Rename `stores` to `storage`
Without `Stores` in there, the old name no longer really fit.
