+++ crates/fj-kernel/src/objects/mod.rs
@@ -70,7 +70,7 @@
 //! As of this writing, most objects are not managed in the centralized object
 //! storage. Changing this is an ongoing effort ([#1021]).
 //!
-//! [`Handle`]: crate::stores::Handle
+//! [`Handle`]: crate::storage::Handle
 //! [#1021]: https://github.com/hannobraun/Fornjot/issues/1021
 
 mod curve;
@@ -83,7 +83,7 @@ mod solid;
 mod surface;
 mod vertex;
 
-use crate::stores::Store;
+use crate::storage::Store;
 
 pub use self::{
     curve::{Curve, GlobalCurve},
