+++ crates/fj-kernel/src/partial/objects/vertex.rs
@@ -3,7 +3,7 @@ use fj_math::Point;
 use crate::{
     objects::{Curve, GlobalVertex, Stores, Surface, SurfaceVertex, Vertex},
     partial::{HasPartial, MaybePartial},
-    stores::Handle,
+    storage::Handle,
 };
 
 /// A partial [`Vertex`]
