+++ crates/fj-kernel/src/storage/handle.rs
